+++ sechub-adapter-pds/src/test/java/com/mercedesbenz/sechub/adapter/pds/PDSWiremockTestSupport.java
@@ -12,8 +12,8 @@ import java.util.UUID;
 import org.springframework.http.HttpStatus;
 
 import com.github.tomakehurst.wiremock.WireMockServer;
-import com.mercedesbenz.sechub.adapter.pds.data.PDSJobStatus.PDSAdapterJobStatusState;
 import com.mercedesbenz.sechub.commons.model.SecHubMessagesList;
+import com.mercedesbenz.sechub.commons.pds.data.PDSJobStatus.PDSAdapterJobStatusState;
 import com.mercedesbenz.sechub.test.JSONTestUtil;
 import com.mercedesbenz.sechub.test.WiremockUrlHistory;
