+++ sechub-commons-pds/src/main/java/com/mercedesbenz/sechub/commons/pds/data/PDSJobParameterEntry.java
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: MIT
-package com.mercedesbenz.sechub.adapter.pds.data;
+package com.mercedesbenz.sechub.commons.pds.data;
 
 public class PDSJobParameterEntry {
