+++ sechub-integrationtest/src/main/java/com/mercedesbenz/sechub/integrationtest/api/AssertPDSStatus.java
@@ -5,7 +5,7 @@ import static org.junit.Assert.*;
 
 import java.util.UUID;
 
-import com.mercedesbenz.sechub.adapter.pds.data.PDSJobStatus.PDSAdapterJobStatusState;
+import com.mercedesbenz.sechub.commons.pds.data.PDSJobStatus.PDSAdapterJobStatusState;
 import com.mercedesbenz.sechub.integrationtest.internal.IntegrationTestPDSJobStatus;
 
 public class AssertPDSStatus {
